feat: new eventhandling for eventbus
Implemented a simple event emitter that does the same as the eventBus
did before but as an object instead of a static class. The new emitter
also no longer relies on the dom's document to emit the events. The general
storage was adjusted to handle the internal events (watch) using the new
event emitter. The ComponentProxy was updated to provide a more solid
implementation of proxy generation.

BREAKING CHANGE: The eventbus, component proxy and the event
registration functions have been moved to a new location